+++ log-processor/src/test/java/core/log/EnvConfTest.java
@@ -1,14 +1,22 @@
 package core.log;
 
-import core.framework.test.EnvConfValidator;
+import core.framework.test.EnvResourceValidator;
+import core.framework.test.EnvWebValidator;
 import org.junit.Test;
+
+import java.io.IOException;
 
 /**
  * @author neo
  */
 public class EnvConfTest {
     @Test
-    public void validateEnvConf() {
-        new EnvConfValidator().validate();
+    public void validateEnvResource() throws IOException {
+        new EnvResourceValidator().validate();
+    }
+
+    @Test
+    public void validateEnvWeb() throws IOException {
+        new EnvWebValidator().validate();
     }
 }